+++ src/router/Routing.tsx
@@ -1,4 +1,4 @@
-import LandingPage from '@/pages/LandingPage';
+import Landing from '@/pages/Landing';
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 import Path from './Path';
@@ -7,7 +7,7 @@ function Routing() {
   return (
     <Routes>
       {/* <Route path="*" element={<NotFoundPage />} /> */}
-      <Route path={Path.LandingPage} element={<LandingPage />} />
+      <Route path={Path.LandingPage} element={<Landing />} />
     </Routes>
   );
 }
